Refactored rest file locations and names
